+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Changed
+- Section control interfaces now include a JSDoc comment with the tab display name and a `{@link}` reference to the containing tab section interface
+
 ## [1.0.0] - 2026-04-09
 ### Added
 - Initial public release